packagekit: Split TwoColumn components in a separate file
This allows us in the next commit to move the RestartServices component
out to its own file with the goal to shrink `updates.jsx` to a more
reasonable line and component count.
Added benefit is that the two components are now in TypeScript making it
easier to convert `updates.jsx` to TypeScript.
